apps/transcription/requirements.txt
modal
openai-whisper
boto3
pydantic
cryptography